requirements.txt
fastapi[all]
websockets
python-dotenv
